use comparison network to determine "sameness"
All of the other distance metrics I've tried (L1, L2, cosine) caused the network to collapse to always giving the same output every time. Hopefully with this, the comparison network will be the only thing to ever collapse, and then it'll be able to "get back up again" if/when the RNN starts producing useful features.